+++ firenado/core/websocket.py
@@ -23,7 +23,7 @@ import tornado.websocket
 
 from firenado import session
 from firenado.core import template
-from firenado.tornado import JSONError
+from firenado.tornadoweb import JSONError
 
 
 # TODO: We need to create a class to avoid those methods repetition here.
