+++ firenado/examples/skell/components/internal/__init__.py
@@ -14,11 +14,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import firenado.tornado
+import firenado.tornadoweb
 from skell.components.internal import uimodules
 
 
-class SkellInternalComponent(firenado.tornado.TornadoComponent):
+class SkellInternalComponent(firenado.tornadoweb.TornadoComponent):
     """ This component just adds another ui modules package to the application
     ui modules stack.
     """
